requirements.txt
Django==2.1.3
django-widget-tweaks==1.4.1
djangorestframework==3.7.3
pytz==2017.3
python-memcached==1.11.0
djangorestframework-simplejwt == 4.3.0
djoser